Fix: 코드 충돌 수정

+++ src/main/java/com/test/bidon/repository/HashtagRepository.java
@@ -16,5 +16,5 @@ public interface HashtagRepository extends JpaRepository<HashTagEntity, Long> {
     @Query("SELECT h.tag FROM HashTaggingEntity hg " +
            "INNER JOIN hg.hashTag h " + 
            "WHERE hg.reviewBoard.id = :reviewBoardId")
-    List<String> findHashtagsByReviewBoardId(@Param("reviewBoardId") Long reviewBoardId);
+    List<String> findHashtagsByReviewBoardId(@Param("reviewBoardId") Integer reviewBoardId);
 }
